Add csv_reader_get_rowitem.
csv_reader_get_rowitem allows you to get a specific item
of a row by heading.

So for example if you wanted to get the `first_name` of a user at row 0,
you'd use:
  item = csv_reader_get_rowitem(reader, 0, "first_name");

Currently if the csv file contains spaces in the headings then this
won't work. You'd have to use a heading name that contains those spaces.
Will fix that later.